refactor: domain 내 하위 디렉토리 구분
