+++ src/main/java/blackjack/domain/game/CardMachine.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.game;
+
+import blackjack.domain.card.Card;
 
 import java.util.ArrayList;
 import java.util.List;
