+++ src/main/java/blackjack/domain/game/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 public class Player extends Human {
