+++ src/main/java/blackjack/domain/game/Players.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import java.util.*;
 import java.util.stream.Collectors;
