+++ src/main/java/blackjack/domain/game/Winner.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import java.util.ArrayList;
 import java.util.List;
