+++ src/test/java/blackjack/domain/game/DealerTest.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardShape;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
